+++ code/presentation/tutorialmanager.cpp
@@ -192,7 +192,17 @@ void TutorialManager::HandleEvent( EventEnum id, void* pEventData )
         return;
     }
 
-    switch( id )
+    // Switched as int rather than EventEnum: EVENT_LOCATOR + LocatorEvent::X
+    // below is a deliberately valid but unnamed EventEnum value (locator
+    // sub-events are reserved a range starting at EVENT_LOCATOR, see
+    // eventenum.h's EVENT_PLACEHOLDER), not a bug - but Clang's -Wswitch
+    // flags any case value that doesn't match a named enumerator of the
+    // switch's type, which a plain int switch doesn't check. Note this
+    // isn't avoided by casting the case value itself (as this file
+    // previously did) - only the switch's controlling expression type
+    // matters (same fix as the other HandleEvent overrides in this
+    // codebase).
+    switch( static_cast<int>( id ) )
     {
         case EVENT_CARD_COLLECTED:
         {
